Chunking parameters modified
In the loading stage, the main hyperparameter is the chunking strategy. The default chunker is the `SentenceSplitter`, which requests sentence and paragraph bounderies, so the end of chunks will likely be sentence ends. Chunk size controls the maximum number of tokens allowed in a chunk. The default value is 1024, here we change it to 256. Chunk overlap contains how many tokens of overlap there should be between chunks.