Using QDrant VectorDB as storage context
So far, we have used local disk persistence for our embedding vectors. Now, we use QDrant vector database. QDrant runs on a kubernetes cluster in Azure in our setup. We need to set its URL, port, api key and the collection name we want to create. These are in the `config.toml` file. A vector database creates an efficient index, that can quickly find chunks with the most semantic simiarity. However, this effect is only relevant when we have chunks in the range of millions.

We need to change a couple config parameters: the details of Qdrant access is stored in the .env file(that is local to everyone), because that is sensitive information. The collection name is stored in config file however, as that is not sensitive. We also removed PERSIST_DIR, as we are not saving our embedding vectors on disk.

In ingestion.py, first we create a qdrant client that can connect to the running Qdrant instance. Then we create a `QdrantVectorStore` object, and create a storage context from it. Then we pass this storage context to our `VectorStoreIndex`. We no longer need to persist, as our embeddings are automatically persisted in the vector DB.