System prompt is changed to include some moderation
We use a new system prompt where set the context of the chatbot and also ask the mdoel to not use previous knowledge.
This is the most simple form of moderation: just modifying the prompt. We do this by setting the "summary_template" of the query engine, that is used to answer questions.