Create a set of test questions
Here, we use a powerful LLM, GPT-4o, to create test questions for a "weaker" llm, in this case, ollama phi3.

Then we create a RagDataset, that generates questions for each chunk, that can be answered based on the given chunk. It gets the chunks as nodes from the vector store.