Code for running evaluation created
Running this code benchmarks the Ollama langauge model, as it is passed to the query_engine. If you don't have Ollama installed, use llm_gpt4 in the query_engine as well.

The code first loads the LabelledRagDataset that we created in the previous step. THen, it downloads the "RagEvaluatorPack" llama pack. Llama pack contains code built by the community to use with LlamaIndex. I recommend checking it out!

The RagEvaluatorPack has 3 arguments:  the query_engine we will benchmark, the dataset to use for benchmarking, that contains questions, refernce answers, and contexts needed to answer them, and a judge_llm, that is used for scoring.

It calculates the correctness, relevancy, faithfulness and semantic similarity scores.